requirements.txt
fastapi
pydantic
SQLAlchemy
uvicorn
pymysql
psycopg2-binary
python-dotenv
pylint
pytest
httpx
pytest-asyncio
coverage
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
pydantic-settings
PyYAML